@@ -77,20 +84,38 @@
       "required": false,
       "title": "AVIF Encoder Path",
       "description": "Path to avifenc command (optional)",
-      "placeholder": "/opt/homebrew/bin/avifenc",
-      "default": "/opt/homebrew/bin/avifenc"
+      "placeholder": "/opt/homebrew/bin/avifenc"
+    },
+    {
+      "name": "avifQuality",
+      "type": "textfield",
+      "required": false,
+      "title": "AVIF Quality",
+      "description": "Quality setting for AVIF conversion (0-100, default: 80)",
+      "placeholder": "80",
+      "default": "80"
+    },
+    {
+      "name": "generateMarkdown",
+      "type": "checkbox",
+      "required": false,
+      "title": "Generate Markdown",
+      "description": "Generate Markdown formatted links instead of plain URLs",
+      "label": "Generate Markdown",
+      "default": false
     }
   ],
   "dependencies": {
     "@aws-sdk/client-s3": "^3.864.0",
     "@raycast/api": "^1.102.4",
-    "@raycast/utils": "^1.17.0"
+    "@raycast/utils": "^1.17.0",
+    "dayjs": "^1.11.18"
   },
   "devDependencies": {
     "@raycast/eslint-config": "^2.0.4",
     "@types/node": "22.13.10",
     "@types/react": "19.0.10",
-    "eslint": "^9.22.0",
+    "eslint": "^9.34.0",
     "prettier": "^3.5.3",
     "typescript": "^5.8.2"
   },
